// src/types/vehicle.ts
export interface Vehicle {
  id: string;
  name: string;
  brand: string;
  model: string;
  year: number;
  batteryCapacity: number;
  consumption: number;
  socketType: string;
  maxChargingPower: number;
  regenEfficiency: number;
  weight: number;
  dragCoefficient: number;
  frontalArea: number;
  rollingResistance: number;
  motorEfficiency: number;
  batteryEfficiency: number;
  thermalEfficiency: number;
  regenBrakingEfficiency: number;
  batteryDegradationFactor: number;
  temperatureEfficiencyFactor: number;
  speedEfficiencyFactor: number;
  loadEfficiencyFactor: number;
  elevationEfficiencyFactor: number;
  regenEfficiencyFactor: number;
  batteryManagementEfficiency: number;
  chargingEfficiency: number;
  auxiliaryPowerConsumption: number;
  climateControlEfficiency: number;
  batteryHeatingEfficiency: number;
  batteryCoolingEfficiency: number;
  inverterEfficiency: number;
  motorControllerEfficiency: number;
  transmissionEfficiency: number;
  wheelEfficiency: number;
  aerodynamicEfficiency: number;
  rollingEfficiency: number;
  batteryAgingFactor: number;
  temperatureImpactFactor: number;
  speedImpactFactor: number;
  loadImpactFactor: number;
  elevationImpactFactor: number;
  regenImpactFactor: number;
  batteryManagementImpactFactor: number;
  chargingImpactFactor: number;
  auxiliaryImpactFactor: number;
  climateControlImpactFactor: number;
  batteryHeatingImpactFactor: number;
  batteryCoolingImpactFactor: number;
  inverterImpactFactor: number;
  motorControllerImpactFactor: number;
  transmissionImpactFactor: number;
  wheelImpactFactor: number;
  aerodynamicImpactFactor: number;
  rollingImpactFactor: number;
} 